+++ frontend/app/speech/speech-client.d.ts
@@ -1,0 +1,9 @@
+import { FC } from 'react';
+
+interface SpeechClientProps {
+  language: string;
+  level: string;
+}
+
+declare const SpeechClient: FC<SpeechClientProps>;
+export default SpeechClient;